Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    int num1, num2, num3;
-    scanf("%d %d %d", &num1, &num2, &num3);
-    printf("Average: %d\n", num1+num2+num3);
+    float num1, num2, num3;
+    scanf("%f %f %f", &num1, &num2, &num3);
+    printf("Average: %f\n", num1+num2+num3);
     return 0;
 }
